refactor: Move ConversationDocument into its own module
